+++ apps/examples/src/examples/HelloWorld.tsx
@@ -4,7 +4,7 @@ import { Description } from "r3f-stage"
 import { Mul, Time } from "shader-composer"
 import { useUniformUnit } from "shader-composer-r3f"
 import { Color } from "three"
-import { composed } from "./lib/composed"
+import { composed } from "@material-composer/patch-material"
 
 export default function HelloWorld() {
   const controls = useControls({ mix: { value: 0.5, min: 0, max: 1 } })
@@ -16,7 +16,7 @@ export default function HelloWorld() {
         <sphereGeometry />
 
         <composed.MeshStandardMaterial>
-          <Modules.Color color={Mul(new Color("hotpink"), Time())} />
+          <Modules.Color color={Mul(new Color("blue"), Time())} />
         </composed.MeshStandardMaterial>
       </mesh>
