+++ apps/examples/src/examples/Playground.tsx
@@ -1,9 +1,11 @@
-import { PatchedMaterialMaster } from "@material-composer/patch-material"
+import {
+  patched,
+  PatchedMaterialMaster
+} from "@material-composer/patch-material"
 import { useControls } from "leva"
 import { Lerp, NormalizePlusMinusOne, Sin, Time } from "shader-composer"
 import { useShader } from "shader-composer-r3f"
 import { Color } from "three"
-import { patched } from "./lib/patched"
 
 export default function Playground() {
   useControls({ foo: { value: 1, min: 0, max: 2 } })
