+++ packages/patch-material/package.json
@@ -19,12 +19,14 @@
   "license": "MIT",
   "sideEffects": false,
   "devDependencies": {
+    "@react-three/fiber": "^8.6.0",
     "@types/three": "^0.143.1",
     "shader-composer": "0.3.1",
     "three": "^0.143.0",
     "typescript": "^4.8.2"
   },
   "peerDependencies": {
+    "@react-three/fiber": "^8.6.0",
     "shader-composer": "^0.3.1",
     "three": ">=0.141.0"
   },
