+++ packages/patch-material/src/composed.tsx
@@ -1,17 +1,17 @@
-import { PatchedMaterialMaster } from "@material-composer/patch-material"
 import { initialModuleState, Module, pipeModules } from "material-composer"
 import {
   ModuleRegistrationContext,
   provideModuleRegistration
 } from "material-composer-r3f/src/moduleRegistration"
-import { DependencyList, useMemo } from "react"
+import React, { DependencyList, useMemo } from "react"
 import { useShader } from "shader-composer-r3f"
+import { PatchedMaterialMaster } from "./master"
 import { patched } from "./patched"
 
 const hasKey = <T extends object>(obj: T, k: keyof any): k is keyof T =>
   k in obj
 
-const useModules = (modules: Module[], deps?: DependencyList) => {
+export const useModules = (modules: Module[], deps?: DependencyList) => {
   const root = useMemo(() => {
     /* Transform state with given modules. */
     const { color, ...state } = pipeModules(
