+++ packages/patch-material/src/index.ts
@@ -1,2 +1,4 @@
+export * from "./composed"
 export * from "./master"
+export * from "./patched"
 export * from "./patchMaterial"
