+++ packages/patch-material/src/lib/useManagedInstance.tsx
